Moved base runtime into its own file
