+++ core/src/runtime/async.rs
@@ -1,0 +1,3 @@
+pub struct Mut {}
+
+pub struct AsyncRuntime {}
